backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==5.0.0
PyJWT==2.8.0
Werkzeug==3.0.3
groq==0.9.0
python-dotenv==1.0.0
